Corrected double in MC DOF and M_PI in CalculateColorIndex()

+++ mandelbulber2/opencl/engines/monte_carlo_dof.cl
@@ -38,8 +38,8 @@ inline void MonteCarloDOF(float3 *startRay, float3 *viewVector, __constant sClIn
 	matrix33 mRot, int *randomSeed)
 {
 #ifdef PERSP_THREE_POINT
-	double randR = 0.0015f * consts->params.DOFRadius * consts->params.DOFFocus
-								 * sqrt(Random(65536, randomSeed) / 65536.0f);
+	float randR = 0.0015f * consts->params.DOFRadius * consts->params.DOFFocus
+								* sqrt(Random(65536, randomSeed) / 65536.0f);
 	float randAngle = Random(65536, randomSeed);
 	float3 randVector = (float3){randR * sin(randAngle), 0.0f, randR * cos(randAngle)};
 	float3 randVectorRot = Matrix33MulFloat3(mRot, randVector);
@@ -49,8 +49,8 @@ inline void MonteCarloDOF(float3 *startRay, float3 *viewVector, __constant sClIn
 	*startRay = consts->params.camera + randVectorRot;
 #else
 	float3 viewVectorTemp = *viewVector;
-	float randR = 0.0015 * consts->params.DOFRadius * consts->params.DOFFocus
-								* sqrt(Random(65536, randomSeed) / 65536.0);
+	float randR = 0.0015f * consts->params.DOFRadius * consts->params.DOFFocus
+								* sqrt(Random(65536, randomSeed) / 65536.0f);
 	float randAngle = Random(65536, randomSeed);
 	float3 randVector = (float3){randR * sin(randAngle), 0.0f, randR * cos(randAngle)};
